Update the RBGateway tests to better check HTTP requests and responses.
This updates the RBGateway test suite to make fewer assumptions about
the API requests and results, which were masking some problems, and as
part of a series of cleanups to prepare for better type safety checks.

All HTTP requests are now checked for all arguments, ensuring we're
passing all the data we expected (and this helped highlight a flaw in
the unit tests where a hosting account URL was not shown).

We now spy on `http_request` and not methods like `http_post`, which
will allow for future type safety checks to kick in.

We also now better check the commits and other information we get back,
ensuring that all the expected data has been parsed from the payload.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9769/