Don't update syntax highlighting prefs when it is disabled globally
When syntax highlighting was disabled globally, we were deleting the
form field from the `AccountSettingsForm`. However, later on we were
trying to read from this field, causing an exception and the my account
page to not load.

Now when syntax highlighting is disabled, we mark the field as disabled
and refuse updates to it. It will also always render as if the user had
it disabled so it does not appear in a disabled but always on state for
the user.

Testing Done:
Ran unit tests.

The my account page no longer raises an exception.

Reviewed at https://reviews.reviewboard.org/r/9796/