Add a long description for the package.
This adds a long description that will show up on PyPI. It uses the
`README.rst` for the repository. This will help visitors on PyPI get an
idea about Review Board, support options, etc. without having to visit
the site or the GitHub repository.

Testing Done:
Built the package and inspected the metadata to make sure the description
was there.

Reviewed at https://reviews.reviewboard.org/r/9804/