Add documentation on configuring Mattermost.
This adds new documentation on configuring the Mattermost integration.
This heavily mirrors the new Slack documentation, which uses many of the
same bits of UI, and differs only in naming and how to generate an
Incoming WebHook.

Testing Done:
Built the docs, checked for typos and image inconsistencies.

Reviewed at https://reviews.reviewboard.org/r/9803/