Add documentation on configuring the I Done This integration.
This covers the new I Done This integration, its purpose, how to set up
an integration configuration, and how to locate and specify API Tokens
for each user that wants to post. It also updates the documentation for
the My Account page to specify the I Done This section (which is only
shown when a configuration is enabled).

Testing Done:
Built the docs. Checked for typos, image inconsistencies, and links.

Reviewed at https://reviews.reviewboard.org/r/9812/