Fix a JavaScript breakage from the new Gerrit instructions.
Just prior to 3.0.4, new instructions were added for the Gerrit hosting
service, pointing to the installation docs. This resulted in a bad call
to `interpolate()`, breaking the JavaScript on the page. This change
fixes that bad call.

I've also updated it to open instructions in a new tab, so people's
forms don't get erased.

Testing Done:
Checked the page. No more JavaScript errors, and the instructions and
link worked (and opened in a new tab).

Reviewed at https://reviews.reviewboard.org/r/9820/