Fix issues with saving custom API token policies.
Even after saving an API token policy, the action would still show that
it's set to "Full access," making it seem like the custom policy didn't
save. If clicked again, this could override the custom policy. This was
happening because we never actually set the policy type to "custom."

This change does that, setting the type when saving the custom policy.

Testing Done:
Selected Custom and edited the policy. Clicked Save. Saw that it was set
to Custom, and that reloading or editing it again showed my proper policy.

Bugs closed: 4663

Reviewed at https://reviews.reviewboard.org/r/9823/