Correctly check for search backend validation in search settings form
The search settings form was assuming that the `search_backend_id` field
would be available in `clean`, but the value is only present if the
backend passes validation. We now check for its existence instead of
assuming that it does exist and carry on with validation accordingly.

Testing Done:
- Ran unit tests.
- Selecting "elasticsearch" on the search settings form without it
  installed no longer results in a 500.
- Enabling Elasticsearch works.
- Enabling Whoosh works.

Bugs closed: 4668

Reviewed at https://reviews.reviewboard.org/r/9824/