Check the elasticsearch version for compatibility during validation.
The Elasticsearch backend checks to make sure the `elasticsearch` module
is installed, but doesn't check its version. Haystack only supports the
1.x and 2.x series, and people these days are going to get much newer
versions than that. We now check the version of the module, and provide
an error message stating the version requirements of both the module and
the service.

Testing Done:
Tested without the module, with a 6.x version of the module, and a 2.x
version. I saw the error in the first two conditions, and validation
passed in the last.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9829/