Explicitly use a UTC timezone for Bazaar.
When working with Bazaar, we normalize all date-based revisions to UTC.
This works so long as `settings.TIME_ZONE` is `UTC`, but if it's
anything else, revision lookup will fail. This is because the `TZ`
environment variable gets set to the value in `settings.TIME_ZONE` by
Django, and this gets inherited by Bazaar, causing all variable lookups
to be offset further by the timezone.

By forcing `TZ=UTC`, we can ensure that Bazaar will be able to match up
any revisions we throw its way.

Unit tests were updated to check for this case.

Testing Done:
Tested that the unit tests failed without the fix both for the new tests
and for existing ones with `settings.TIME_ZONE = 'US/Pacific'`.

Verified all tests are fixed with this change, with both `UTC` and
`US/Pacific` set in `settings_local.py`.

Reviewed at https://reviews.reviewboard.org/r/9836/