Fix timezone-related issues with looking up Mercurial commits.
When loading information from Mercurial commits, we assume that server
timezone is set to UTC. If it's anything else, the commit will be
converted to the wrong time. This was due to the usage of
`datetime.fromtimestamp`, which was converting into the local timezone.
We now use `datetime.utcfromtimestamp` to explicitly convert to UTC.

A unit test was added to test this situation.

Testing Done:
Tested the original issue with a non-UTC timezone set, and saw the
failures in an existing commit. Also tested that the new unit test also
fails.

Verified after the commit that all unit tests pass, no matter my
configured timezone.

Reviewed at https://reviews.reviewboard.org/r/9837/