Fix regressions in auto-sizing of text fields for the review request page.
The text fields (Description and Testing Done) on the review request
page are meant to grow as you add content and shrink as you remove it.

This regressed in 3.0 with the field rework due to the "resize" event
not making it up to the field. `RB.TextEditorView` was emitting the
event, but `RB.InlineEditorView` wasn't passing it up, and neither was
`RB.ReviewRequestFields.TextFieldView`.

Chaining these events up to `TextFieldView` ensures that layout
calculation will take place as content is added to or removed from the
text field, restoring the ability to auto-size the field.

Testing Done:
Manually tested that this fixed issues with auto-resizing the text fields
on RBCommons.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9853/