Link to LocalSite user pages from LocalSite datagrids
A change a while back upgraded the UsernameColumn to use the new
`link_func` functionality of Djblets datagrids. However, it regressed
the created URL to the absolute URL of the user (which always lives in
the global namespace) instead of the per-LocalSite URL when viewing a
datagrid on a LocalSite. This change fixes that issue.

Testing Done:
Ran unit tests.
Viewed a datagrid on a LocalSite and saw the correct URL for the
username column.

Reviewed at https://reviews.reviewboard.org/r/9855/