Check for enabled features on local sites when on a global site
When browsing a non-LocalSite specific page, if any LocalSite the user is
a member of has a feature enabled then it should be enabled for that
user while browsing the non-LocalSite page. Previously, this was not
taken into account and features that a user should otherwise have access
to would be disabled if they were enabled only for their LocalSites.
This has been corrected.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9705/