Fix striking out the Ship It text when revoking.
When revoking Ship It, the text is supposed to be updated to show a
strike-through effect. This broke at some point when the template for
the review entry was changed to better group some parts of a comment
thread, preventing the element from being updated.

This change fixes the unit tests to reflect the current structure, and
fixes the query to look up the right elements.

Testing Done:
Revoked a Ship It. Saw that the text updated.

Unit tests all pass.

Reviewed at https://reviews.reviewboard.org/r/9862/