Work around errors in patch bundle URL resolution.
Web crawlers have a nasty habit of trying to index diff fragments
(somehow), and don't always pass in all the right data in the URL. When
we encounter the unsurprising errors, we'd try to show a link to the
patch bundle, but reversing this URL would fail.

This change works around that by catching the `NoReverseMatch` and just
including an empty link instead.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9917/