Ensure extra_data is set before modifying it in the API.
When working with extra_data in the API, we assumed the version on the
object was a valid dictionary. If this was instead `None`, API requests
would fail. This could happen with very old objects, ones explicitly set
to `None` through a script/terminal, or if something went wrong when
setting up the object.

We now check for this and set it to `{}` before operating on it if
necessary.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9907/