Add PII-safe titles for the user pages.
This adds a context variable for the user pages that provide PII-safe
titles for use in services like Google Analytics. These titles are
stripped of usernames, and just provide a generic page title to go with
what will be a generic URL.

Testing Done:
Manually tested that the Google Analytics template was passing in the
PII-safe titles for these pages.

Reviewed at https://reviews.reviewboard.org/r/9944/