Add registration of built-in privacy consent requirements.
This registers the built-in Gravatar consent rquirement, which will show
up on the My Account page. This is the only consent requirement needed
directly by Review Board. It also includes a custom-made icon for
Gravatar.

This change was meant to go up and land before the previous changes for
the decorator redirects. It fixes test breakages introduced in that
change.

Testing Done:
Unit tests pass.

Manually tested the consent requirement and its appearance in the My
Account page.

Reviewed at https://reviews.reviewboard.org/r/9953/