Add acceptance of policies during user sign-up.
This change adds a new checkbox to the account registration field to
allow users to indicate their acceptance of the linked privacy policy
and/or terms of service. This is shown when URLs for these are
configured in the privacy section of the admin settings and user consent
is turned on. When saved, the form will then create the user and record
their acceptance of the policies using the consent requirement.

Testing done:
Verified the display of the form with various configurations (consent
turned on/off, policies linked or not). Saw that when the field was
visible, it was required, and creating a user would then immediately
redirect to the privacy consent page with the policies marked as
accepted.

Reviewed at https://reviews.reviewboard.org/r/9962/