Lazily mark Gravatar requirement intent as safe
Previously we were calling `mark_safe` on the reult of `ugettext_lazy`
in the `GravatarConsentRequirement`, which causes the string to be
translated at that time. This would result in every user (regardless of
language) receiving the string in whatever the site-wide langauge is,
which is not ideal.

We now use `mark_safe_lazy` so that translation occurs when it is
rendered in a template.

Reviewed at https://reviews.reviewboard.org/r/9975/