Render only consent form when user has pending consent decisions
When a user has pending consent decisions or have not accepted the policy
requirements (if set), the My Account page will only render the privacy
form. This is to prevent users from navigating to another page and
failing to complete the form.

Additionally, if a user is redirected to this form to complete consent
decisions, they will be redirected to the original page once they have
completed them.

Testing Done:
- Ran unit tests.
- Manually tested form submission and was redirected correctly.

Reviewed at https://reviews.reviewboard.org/r/9974/