Do not redirect to consent page from a grid-only datagrid
Datagrids periodically make requests to update themselves and replace
their content with the response content. If a user has datagrid open and
their consent decisions become invalid (i.e., they declined the privacy
policy in another tab or a new requirement was added), the next reload
of the datagrid will result in a redirect, causing the consent form to
be rendered inside the datagrid itself.

To correct for this situation, the `valid_prefs_required` now accepts a
`disable_consent_checks` argument which is a function that determines if
consent checks should be skipped for the current request. When this
argument is not provided, the decorator behaves as it did originally.
However, when it is provided, this check will be taken into account.

All datagrid views have been updated to use this new feature so that we
do not end up embedding the consent form in any datagrid.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9980/