Fix a handful of test docstrings for the webhooks API.
We had a sizable number of unit tests for the webhooks API that had
non-conforming docstrings. These left out the HTTP method and URL,
making it harder to tell that the tests applied to the API and not to
webhooks in general. This change updates all the unit tests to fix up
those docstrings.

Testing Done:
Ran the webhook API unit tests. Saw consistency with the docstrings
in the test output.

Reviewed at https://reviews.reviewboard.org/r/10028/