Separate the ReviewRequestPageDataMixin from BuiltinFieldMixin
The `ReviewRequestPageDataMixin` was built on the `BuiltinFieldMixin`.
However, this requires that the field being represented be available on
the review request draft as `field_id` or else the mixin will use the
review request instead. This is not ideal when the represented field is
derived state or not directly accessible as an attribute/property on the
review request draft, e.g., it is an attribute of the draft diff
instead.

Now the two are more separate and fields that require the old behaviour
inherit from both.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10082/