Use --interactive flag for PostgreSQL 9.2+
Since PostgreSQL 9.2, the --interactive flag to createuser is required
to get the "ask me for everything I didn't specify" behavior that was
the default behavior in 9.1 and previous [1] and that the rest of this
document assumes.

The --interactive flag is not recognized in 9.1, so this will not work
with 9.1 and previous.  But, PostgreSQL 9.1 was EOL'd 22 months ago
[2], so it's not worth complicating these instructions with an "If your
PostgreSQL version is ..." branch here to support such an old release.

[1] Compare:
    https://www.postgresql.org/docs/9.1/static/app-createuser.html
    https://www.postgresql.org/docs/9.2/static/app-createuser.html
[2] https://www.postgresql.org/support/versioning/

Reviewed at https://reviews.reviewboard.org/r/10111/