Fix visual alignment issues caused by move detection flags.
The move detection flags alter both the alignment of the line numbers
and code, which can cause some confusion during review. Much of this was
due to some old CSS that impacted the line number cell, and there not
being enough room for the move range line in the code cell.

The CSS has been fixed to address all this. We still show a 4px move
range line, but now it's spread across the line number header and code
cell, 3px in the header and 1px in the cell. This takes advantage of the
spacing we already have available in both cells without worrying about
bumping right up against any text, or having to increase the default
padding in either cell.

Testing Done:
Tested in Chrome and Firefox. Checked for perfect pixel alignment in the
cells.

Reviewed at https://reviews.reviewboard.org/r/10077/