Display commit lists for review requests created with history
The `CommitListField` renders the summaries (and authors if they differ
from the review request submitter) of the commits that make up a
multi-commit review request. This field has been added to a new
`ExtraFieldSet`, which exists below the `MainFieldSet`. This is due to
the `RB.ReviewRequestEditorView` expecting the last field in the main
field set to be an editable text field and working around this
limitation was more complex and convoluted than adding a new fieldset.

The JS `RB.ReviewRequestFields.CommitListFieldView` is responsible for
toggling between the full commit message (if it is longer than a single
line) and the summary.

Testing Done:
- Ran unit tests.
- Ran JS tests.
- Uploaded review requests with multiple commits and saw the correct
  commits displayed in the field and in change descriptions.
- Uploaded a plain review request and did not see the new field on
  the review request page or change description box.

Reviewed at https://reviews.reviewboard.org/r/10094/