Compute pre-patch files for FileDiffs in commit histories
We can now generate the pre-patch version of a file that is modified
multiple times in a commit history. This allows the diffviewer to show
side-by-side diffs in these cases, however this behaviour is not correct
as it will show files with the same name multiple times when it should
show a condensed diff of the entire history. This will be addressed in a
later patch.

This also allows the original and patched file API endpoints to work for
files modified multiple times in a commit history.

Testing Done:
- Ran unit tests.
- Was able to view diffs in the diffviewer that depended on previous
  FileDiffs.

Reviewed at https://reviews.reviewboard.org/r/10119/