Show cumulative diffs in the diffviewer
Previously, `FileDiffs` for review requests with history would not show
cumulatively. That is, you would see only the result of applying that
`FileDiff`, not the result of applying the entire stack of `FileDiffs`
in the history graph.

We now compute the entire file history, including deletions, and use
that to compute the cumulative diff. In other words, for a review
request with commits `A -> B -> C` where all three commits modify the
same file, previously you would see the result of `A`, `B`, and `C`
independently. Now the diffviewer shows `A`, `A..B`, and `A..C`.

Files are still repeated (as indicated above) so this is not yet a true
cumulative diff. A future patch will update the diffviewer to only show
the leaf `FileDiffs` in the history graph.

Testing Done:
- Ran unit tests.
- Uploaded a multi-commit review request and looked at the diff. I
  observed cumulative (but repeated) diffs.

Reviewed at https://reviews.reviewboard.org/r/10123/