Exclude ancestor FileDiffs from the diffviewer
When a review request is created with commit history, it is likely that
there will be `FileDiff`s in multiple commits that modify the same file.
For any `FileDiff` in a commit, all `FileDiff`s in previous commits are
known as "ancestor" `FileDiff`s.

These ancestor `FileDiff`s are now excluded from the diffviewer. This
removes the duplication of seeing the same file listed multiple times
with incremental changes. However, currently only the latest change for
that file is shown. True cumulative diffs are supported in /r/10130/.

Testing Done:
- Ran unit tests.
- Viewed a review request with commit history and did not see repeated
  files.

Reviewed at https://reviews.reviewboard.org/r/10124/